refactor(test): move inspect CLI surface tests to tests/cli/
Address PR review comments on #676:

1. Move TestInspectCliSurface (11 tests) and TestInspectListTasks (6 tests)
   from tests/e2e/ to tests/cli/test_inspect_cli.py per tests/CLAUDE.md:
   CLI surface tests belong under tests/cli/, not tests/e2e/.
   They now run under the default CI filter without -m e2e.

2. Replace tautological assertions in test_list_tasks_all_lines_are_strings
   (isinstance(str.strip(), str) is always true; len > 0 is guaranteed by
   the if-guard) with a regex invariant: each task line must match
   r"^[a-z][a-z0-9-]*$".

3. Collapse two redundant tests (test_list_tasks_no_model_flag_needed and
   test_list_tasks_overrides_missing_model_requirement both just asserted
   exit_code == 0, already covered by test_list_tasks_exits_zero) into a
   single test_list_tasks_is_sorted that verifies the --list-tasks output
   is in ascending lexicographic order.
